+++ MercSkinsFix/MercSkinsFixPlugin.cs
@@ -1,11 +1,16 @@
 using BepInEx;
+using BepInEx.Bootstrap;
 using BepInEx.Configuration;
+using HarmonyLib;
+using MonoMod.RuntimeDetour;
 using RiskOfOptions;
 using RiskOfOptions.Options;
 using RoR2;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Reflection;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 
@@ -17,9 +22,16 @@ namespace MercSkinsFix
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "Gorakh";
         public const string PluginName = "MercSkinsFix";
-        public const string PluginVersion = "1.0.0";
+        public const string PluginVersion = "1.1.0";
 
         internal static MercSkinsFixPlugin Instance { get; private set; }
+
+        static readonly Dictionary<Assembly, BepInEx.PluginInfo> _assemblyPluginLookup = [];
+        static readonly Dictionary<SkinDef, Assembly> _skinOwnerAssemblies = [];
+
+        static readonly Dictionary<Assembly, DateTime?> _assemblyTimestampLookup = [];
+
+        Hook _scriptableObjectCreateInstanceHook;
 
         void Awake()
         {
@@ -28,6 +40,8 @@ namespace MercSkinsFix
             Log.Init(Logger);
 
             Instance = SingletonHelper.Assign(Instance, this);
+
+            _scriptableObjectCreateInstanceHook = new Hook(SymbolExtensions.GetMethodInfo(() => ScriptableObject.CreateInstance(default(Type))), ScriptableObject_CreateInstance);
 
             stopwatch.Stop();
             Log.Info_NoCallerPrefix($"Initialized in {stopwatch.Elapsed.TotalSeconds:F2} seconds");
@@ -38,6 +52,98 @@ namespace MercSkinsFix
         void OnDestroy()
         {
             Instance = SingletonHelper.Unassign(Instance, this);
+
+            _scriptableObjectCreateInstanceHook?.Dispose();
+            _scriptableObjectCreateInstanceHook = null;
+        }
+
+        delegate ScriptableObject orig_ScriptableObject_CreateInstance(Type type);
+        static ScriptableObject ScriptableObject_CreateInstance(orig_ScriptableObject_CreateInstance orig, Type type)
+        {
+            ScriptableObject obj = orig(type);
+
+            if (obj is SkinDef skinDef)
+            {
+                StackTrace stackTrace = new StackTrace();
+
+                for (int i = 0; i < stackTrace.FrameCount; i++)
+                {
+                    StackFrame frame = stackTrace.GetFrame(i);
+
+                    Assembly assembly = frame?.GetMethod()?.DeclaringType?.Assembly;
+                    if (assembly == null || assembly == Assembly.GetExecutingAssembly())
+                        continue;
+
+                    if (!_assemblyPluginLookup.TryGetValue(assembly, out BepInEx.PluginInfo plugin))
+                    {
+                        plugin = null;
+
+                        foreach (BepInEx.PluginInfo pluginInfo in Chainloader.PluginInfos.Values)
+                        {
+                            if (pluginInfo.Instance && pluginInfo.Instance.GetType().Assembly == assembly)
+                            {
+                                plugin = pluginInfo;
+                                break;
+                            }
+                        }
+
+                        _assemblyPluginLookup.Add(assembly, plugin);
+                    }
+
+                    if (plugin == null)
+                        continue;
+
+                    if (!_assemblyTimestampLookup.ContainsKey(assembly))
+                    {
+                        DateTime? assemblyTimestamp = null;
+
+                        string assemblyLocation = assembly.Location;
+                        if (!string.IsNullOrEmpty(assemblyLocation) && File.Exists(assemblyLocation))
+                        {
+                            // https://stackoverflow.com/questions/2050396/getting-the-date-of-a-net-assembly
+
+                            const int peHeaderOffset = 60;
+                            const int linkerTimestampOffset = 8;
+
+                            try
+                            {
+                                using FileStream fs = File.Open(assemblyLocation, FileMode.Open, FileAccess.Read, FileShare.Read);
+
+                                byte[] buffer = new byte[2048];
+                                int bytesRead = fs.Read(buffer, 0, 2048);
+                                if (bytesRead >= peHeaderOffset)
+                                {
+                                    int linkerTimestampLocation = BitConverter.ToInt32(buffer, peHeaderOffset) + linkerTimestampOffset;
+                                    if (bytesRead >= linkerTimestampLocation + sizeof(int))
+                                    {
+                                        int timestampSeconds = BitConverter.ToInt32(buffer, linkerTimestampLocation);
+
+                                        DateTime linkerTimestampUTC = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc).AddSeconds(timestampSeconds);
+
+                                        // Timestamp is not *always* number of seconds
+                                        // TODO: Handle other cases? Filter out for now
+                                        if (linkerTimestampUTC > new DateTime(2000, 1, 1, 0, 0, 0, DateTimeKind.Utc) && linkerTimestampUTC < DateTime.UtcNow)
+                                        {
+                                            assemblyTimestamp = linkerTimestampUTC;
+                                        }
+                                    }
+                                }
+                            }
+                            catch (Exception e)
+                            {
+                                Log.Error_NoCallerPrefix($"Failed to determine assembly date for {assembly.GetName()}: {e}");
+                            }
+                        }
+
+                        _assemblyTimestampLookup.Add(assembly, assemblyTimestamp);
+                    }
+
+                    _skinOwnerAssemblies.Add(skinDef, assembly);
+                    break;
+                }
+            }
+
+            return obj;
         }
 
         void onLoad()
@@ -69,10 +175,17 @@ namespace MercSkinsFix
             const string SETTING_MOD_GUID = PluginGUID;
             const string SETTING_MOD_NAME = "Mercenary Skins Fix";
 
+            Dictionary<string, HashSet<string>> usedSectionKeys = [];
+
+            Config.SaveOnConfigSet = false;
+
             foreach (SkinDef skin in skins)
             {
                 if (Array.IndexOf(vanillaMercSkinDefs, skin) >= 0)
                     continue;
+
+                Assembly ownerAssembly = _skinOwnerAssemblies.GetValueSafe(skin);
+                BepInEx.PluginInfo ownerPlugin = ownerAssembly != null ? _assemblyPluginLookup.GetValueSafe(ownerAssembly) : null;
 
                 void setApplyToSkin(bool apply)
                 {
@@ -114,7 +227,66 @@ namespace MercSkinsFix
                     }
                 }
 
-                ConfigEntry<bool> shouldApplyToSkinConfig = Config.Bind("Skins", $"Fix {Language.GetString(skin.nameToken, "en")} ({skin.name})", false, "Apply the fix to this skin");
+                string sectionName = "Unknown";
+                bool fixEnabledByDefault = false;
+                if (ownerPlugin != null && ownerAssembly != null)
+                {
+                    if (_assemblyTimestampLookup.TryGetValue(ownerAssembly, out DateTime? assemblyTimestamp) && assemblyTimestamp.HasValue)
+                    {
+                        DateTime devotionReleaseDate = new DateTime(2024, 5, 20, 0, 0, 0, DateTimeKind.Utc);
+                        if (assemblyTimestamp.Value < devotionReleaseDate)
+                        {
+                            fixEnabledByDefault = true;
+                        }
+                    }
+
+                    sectionName = $"{ownerPlugin.Metadata.Name} ({System.IO.Path.GetFileName(ownerAssembly.Location)})";
+                }
+
+                string skinName;
+                if (string.IsNullOrEmpty(skin.nameToken))
+                {
+                    if (string.IsNullOrEmpty(skin.name))
+                    {
+                        skinName = skin.skinIndex.ToString();
+                    }
+                    else
+                    {
+                        skinName = skin.name;
+                    }
+                }
+                else
+                {
+                    if (string.IsNullOrEmpty(skin.name))
+                    {
+                        skinName = Language.GetString(skin.nameToken, "en");
+                    }
+                    else
+                    {
+                        skinName = $"{Language.GetString(skin.nameToken, "en")} ({skin.name})";
+                    }
+                }
+
+                if (usedSectionKeys.TryGetValue(sectionName, out HashSet<string> usedKeys))
+                {
+                    if (!usedKeys.Add(skinName))
+                    {
+                        string unmodifiedSkinName = skinName;
+
+                        int i = 1;
+                        do
+                        {
+                            skinName = $"{unmodifiedSkinName} ({i})";
+                            i++;
+                        } while (!usedKeys.Add(skinName));
+                    }
+                }
+                else
+                {
+                    usedSectionKeys.Add(sectionName, [skinName]);
+                }
+
+                ConfigEntry<bool> shouldApplyToSkinConfig = Config.Bind(sectionName, $"Fix {skinName}", fixEnabledByDefault, "Apply the fix to this skin");
 
                 ModSettingsManager.AddOption(new CheckBoxOption(shouldApplyToSkinConfig), SETTING_MOD_GUID, SETTING_MOD_NAME);
 
@@ -132,6 +304,9 @@ namespace MercSkinsFix
                     setApplyToSkin(applyToSkin);
                 };
             }
+
+            Config.SaveOnConfigSet = false;
+            Config.Save();
 
             ModSettingsManager.SetModDescription("Fix broken Mercenary skins after the Devotion update", SETTING_MOD_GUID, SETTING_MOD_NAME);
 
@@ -160,6 +335,9 @@ namespace MercSkinsFix
                     ModSettingsManager.SetModIcon(iconSprite, SETTING_MOD_GUID, SETTING_MOD_NAME);
                 }
             }
+
+            _scriptableObjectCreateInstanceHook?.Dispose();
+            _scriptableObjectCreateInstanceHook = null;
         }
     }
 }