+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**1.1.0 Changes:**
+
+* Fixed getting stuck at 100% load if several Mercenary skins share the same name
+
+* Added automatic detection system to determine with skins need to be fixed (See README for details)
+
 **1.0.0 Changes:**
 
 * First release